+++ test/extended_math_test.dart
@@ -12,11 +12,18 @@ void main() {
   final v2 = Vector(<double>[-2, 5, 1]);
 
   // [[0.5, -0.0, 0.0], [-0.0, 0.125, -0.0], [0.0, -0.0, 0.2]]
-  //print(v1.inverse().data);
+  // print(v1.inverse().data);
 
   final m = SquareMatrix(<List<double>>[
-    <double>[1, 0],
-    <double>[6, 1]
+    <double>[1, 2, 1, 13],
+    <double>[4, 3, -2, 56],
+    <double>[3, 2, 1, 33],
+    <double>[7, 9, 15, 11]
   ]);
-  print(m.isOrthogonal());
+  final m1 = SquareMatrix(<List<double>>[
+    <double>[2, 4],
+    <double>[5, 3]
+  ]);
+  print(m.eliminate(<double>[8, 4, 10, 13]).data);
+  // print(m1.insertRow(<double>[1, 2], index: 1).data);
 }